+++ internal/eqemuanalytics/releases.go
@@ -3,6 +3,7 @@ package eqemuanalytics
 import (
 	"context"
 	"github.com/google/go-github/v41/github"
+	gocache "github.com/patrickmn/go-cache"
 	"golang.org/x/oauth2"
 	"net/http"
 	"os"
@@ -11,6 +12,7 @@ import (
 
 type Releases struct {
 	client *github.Client
+	cache  *gocache.Cache
 }
 
 func NewReleases() *Releases {
@@ -28,10 +30,15 @@ func NewReleases() *Releases {
 		client = github.NewClient(tc)
 	}
 
-	return &Releases{client: client}
+	return &Releases{client: client, cache: gocache.New(5*time.Minute, 10*time.Minute)}
 }
 
 func (r *Releases) getReleases() ([]*github.RepositoryRelease, error) {
+	// return cached releases if they exist
+	if x, found := r.cache.Get("releases"); found {
+		return x.([]*github.RepositoryRelease), nil
+	}
+
 	releases, _, err := r.client.Repositories.ListReleases(
 		context.Background(),
 		"EQEmu",
@@ -44,6 +51,9 @@ func (r *Releases) getReleases() ([]*github.RepositoryRelease, error) {
 	if err != nil {
 		return nil, err
 	}
+
+	// cache releases
+	r.cache.Set("releases", releases, time.Minute*1)
 
 	return releases, nil
 }